added coordinates
I added the coordinates of the video "The Finnish-Russian Border, and a Conveniently Placed Hut" where the bench is. I tried to find with this very map, but it wasn't listed. So I did the hard work of finding the bench, and wanted to list it for future people. Via street view, it is the same bench with the quirky situation of being behind a sign but not in the buffer zone.